dowellpermutation/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.10.0
click==8.1.3
colorama==0.4.6
cssbeautifier==1.14.7
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
djlint==1.19.3
EditorConfig==0.12.3
html-tag-names==0.1.2
html-void-elements==0.1.0
importlib-metadata==5.0.0
jsbeautifier==1.14.7
mypy-extensions==0.4.3
numpy==1.23.4
pandas==1.5.1
pathspec==0.10.1
platformdirs==2.5.2
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
regex==2022.10.31
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
zipp==3.10.0
